image_steganography/requirements.txt
PyQt5
Pillow
numpy